use logging to inform of failed LRO
